refactor: decouple producers.py
