+++ .github/workflows/test.yml
@@ -3,7 +3,7 @@ on:
   pull_request:
   push:
     branches:
-      - main
+      - decouple_producers
 
 jobs:
   datafusion_producer:
