+++ substrait_consumer/functional/queries/sql/comparison_functions_sql.py
@@ -1,4 +1,6 @@
-from substrait_consumer.producers import *
+from substrait_consumer.producers.datafusion_producer import DataFusionProducer
+from substrait_consumer.producers.duckdb_producer import DuckDBProducer
+from substrait_consumer.producers.isthmus_producer import IsthmusProducer
 
 SQL_SCALAR = {
     "not_equal": (
